requirements.txt
azure-functions
python-docx
